requirements.txt
python-telegram-bot
whisper
openai
python-dotenv
